requirements.txt
pandaSDMX
pandas-stubs
pywikibot
